requirements.txt
djangorestframework==3.13.1
pymongo[srv]==4.1.0 
django-cors-headers==3.11.0
requests==2.20.0
python-decouple==3.6
bcrypt==3.2.0